Refactored much of the active days logic so that it's not all clustered in a single class

+++ SHModels/SHModels/Models/SHDaily/SHDaily.h
@@ -19,7 +19,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface SHDaily : NSManagedObject<SHDueDateItemProtocol, SHTitleProtocol>
 @property (strong,nonatomic) SHDailyActiveDays *activeDaysContainer;
-@property (readonly,nonatomic) int32_t rate;
+@property (readonly,nonatomic) NSInteger rate;
 @property (readonly,nonatomic) BOOL isCompleted;
 @property (strong,nonatomic) NSObject<SHDateProviderProtocol> *dateProvider;
 -(void)setupInitialState;
